+++ extensions/aiven/CHANGELOG.md
@@ -1,3 +1,9 @@
 # Aiven Changelog
 
+## [Initial Version] - {PR_MERGE_DATE}
+
+- Add "Open in Browser" `Action`s
+- Add `tooltip`s to service items
+- View Service **Backups**
+
 ## [Initial Version] - 2025-01-09